Django_DRF/requirements.txt
asgiref==3.7.2
Django==5.0.4
djangorestframework==3.15.1
setuptools==68.1.2
sqlparse==0.4.4
